Moved Auth from component to organisms
