+++ components/organisms/Auth/Onboarding/Step2/OnboardingStep2.tsx
