+++ components/organisms/Auth/Onboarding/Step4/OnboardingStep4.tsx
