+++ components/organisms/Auth/Onboarding/Step6/OnboardingStep6.tsx
