+++ components/organisms/Auth/Onboarding/Step8/OnboardingStep8.tsx
