+++ components/organisms/Auth/Onboarding/_DEPRECATED_Step1/OnboardingStep1.tsx
